% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{mason}
\alias{mason}
\title{Mason}
\usage{
mason(target, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{target}{\code{id} of target \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{session}{A valid shiny session.}
}
\description{
Mason a rendered \code{\link[=masonryGrid]{masonryGrid()}}.
}
