% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{masonry}
\alias{masonry}
\title{Row}
\usage{
masonry(..., options = list(), id = NULL, classes = "")
}
\arguments{
\item{...}{Any number of \code{\link[=masonryRow]{masonryRow()}}.}

\item{options}{\code{\link[=list]{list()}} of options.}

\item{id}{CSS ID of masonry element.}

\item{classes}{Any additional classes.}
}
\description{
Masonry row
}
