% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonry_add_item}
\alias{masonry_add_item}
\title{Add an item}
\usage{
masonry_add_item(
  target,
  row_index,
  item,
  position = c("start", "end"),
  classes = "",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{target}{\code{id} of target \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{row_index}{Index of row, an integer.}

\item{item}{htmltools \code{tags}, the content of the item.}

\item{position}{Whether to add the new row at the \code{start}
or the \code{end}.}

\item{classes}{Additional classes to add to the row.}

\item{session}{A valid shiny session.}
}
\description{
Add an item to \code{\link[=masonryRow]{masonryRow()}}.
}
