% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonry_remove_item}
\alias{masonry_remove_item}
\title{Remove an item}
\usage{
masonry_remove_item(
  target,
  row_index,
  item_index,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{target}{\code{id} of target \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{row_index}{Index of row, an integer.}

\item{item_index}{index of item to remove.}

\item{session}{A valid shiny session.}
}
\description{
Remove an item from \code{\link[=masonryRow]{masonryRow()}}.
}
