% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonry_remove_row}
\alias{masonry_remove_row}
\title{Remove a row}
\usage{
masonry_remove_row(
  target,
  row_index,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{target}{\code{id} of target \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{row_index}{Index of row, an integer.}

\item{session}{A valid shiny session.}
}
\description{
Remove a row from \code{\link[=masonryGrid]{masonryGrid()}}.
}
