% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{masonryGrid}
\alias{masonryGrid}
\title{Row}
\usage{
masonryGrid(..., styles = list(), id = NULL, classes = "")
}
\arguments{
\item{...}{Any number of \code{\link[=masonryRow]{masonryRow()}}.}

\item{styles}{\code{\link[=list]{list()}} of options for \code{items} and \code{rows},
e.g.: \code{list(items = list(margin = ".5rem"))}.}

\item{id}{CSS ID of masonry element.}

\item{classes}{Any additional classes.}
}
\description{
Masonry row
}
