% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{masonryItem}
\alias{masonryItem}
\title{Item}
\usage{
masonryItem(..., width = NULL)
}
\arguments{
\item{...}{Content of the item.}

\item{width}{Width, a percentage (\code{numeric}, e.g.: \code{.5}).}
}
\description{
Masonry item
}
