% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonryOutput}
\alias{masonryOutput}
\title{Masonry Shiny}
\usage{
masonryOutput(id, styles = list(), classes = "")
}
\arguments{
\item{id}{CSS ID of masonry element.}

\item{styles}{\code{\link[=list]{list()}} of options for \code{items} and \code{rows},
e.g.: \code{list(items = list(margin = ".5rem"))}.}

\item{classes}{Any additional classes.}
}
\description{
Render Masonry grid.
}
