% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{masonryRow}
\alias{masonryRow}
\title{Row}
\usage{
masonryRow(..., min_height = "5rem")
}
\arguments{
\item{...}{Any number of \code{\link[=masonryItem]{masonryItem()}}.}

\item{min_height}{Minimum height of row, a valid CSS value.
This is used because if not set the row goes to 0 height if empty
and users can then no longer drag \code{\link[=masonryItem]{masonryItem()}} to it.}
}
\description{
Masonry row
}
