% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonry_add_item}
\alias{masonry_add_item}
\title{Add an item}
\usage{
masonry_add_item(
  target,
  item,
  row_index = NULL,
  row_id = NULL,
  position = c("start", "end"),
  classes = "",
  session = shiny::getDefaultReactiveDomain(),
  event_id = NULL,
  mason = TRUE,
  id = NULL
)
}
\arguments{
\item{target}{\code{id} of target \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{item}{htmltools \code{tags}, the content of the item.}

\item{row_index, row_id}{Index of id of row.}

\item{position}{Whether to add the new row at the \code{start}
or the \code{end}.}

\item{classes}{Additional classes to add to the row.}

\item{session}{A valid shiny session.}

\item{event_id}{id of event to trigger (\code{input$event_id}) when the item is rendered.}

\item{mason}{Whether to run masonry after adding the item.}

\item{id}{ID of masonry item.}
}
\description{
Add an item to \code{\link[=masonryRow]{masonryRow()}}.
}
