% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonry_get_config}
\alias{masonry_get_config}
\title{Get config}
\usage{
masonry_get_config(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{\code{id} of target \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{session}{A valid shiny session.}
}
\description{
Get config \code{\link[=masonryGrid]{masonryGrid()}}.
}
