% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{masonryGrid}
\alias{masonryGrid}
\title{Row}
\usage{
masonryGrid(..., options = list(), id = NULL, classes = "")
}
\arguments{
\item{...}{Any number of \code{\link[=masonryRow]{masonryRow()}}.}

\item{options}{\code{\link[=list]{list()}} of options.}

\item{id}{CSS ID of masonry element.}

\item{classes}{Any additional classes.}
}
\description{
Masonry row
}
