% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{masonryItem}
\alias{masonryItem}
\title{Item}
\usage{
masonryItem(..., width = 0.2, classes = "")
}
\arguments{
\item{...}{Content of the item.}

\item{width}{Initial width, a percentage (\code{numeric}, e.g.: \code{.5}).}

\item{classes}{Any additional classes.}
}
\description{
Masonry item
}
