% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{masonryOutput}
\alias{masonryOutput}
\alias{renderMasonry}
\title{Masonry Shiny}
\usage{
masonryOutput(id, styles = list(), classes = "")

renderMasonry(expr, styles = list(), env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{ID of masonryGrid.}

\item{styles}{\code{\link[=list]{list()}} of options for \code{items} and \code{rows},
e.g.: \code{list(items = list(margin = ".5rem"))}.}

\item{classes}{Any additional classes.}

\item{expr}{Expression returning \code{\link[=masonryGrid]{masonryGrid()}}.}

\item{env}{Environment to evaluate the \code{expr}.}

\item{quoted}{Whether to quote the expression.}
}
\description{
Render Masonry grid.
}
