% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{masonryRow}
\alias{masonryRow}
\title{Row}
\usage{
masonryRow(..., min_height = "5rem", height = NULL, classes = "")
}
\arguments{
\item{...}{Any number of \code{\link[=masonryItem]{masonryItem()}}.}

\item{min_height}{Minimum height of row, a valid CSS value.
This is used because if not set the row goes to 0 height if empty
and users can then no longer drag \code{\link[=masonryItem]{masonryItem()}} to it.}

\item{height}{Height, fixed height of the row.}

\item{classes}{Any additional classes.}
}
\description{
Masonry row
}
