% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{masonryRow}
\alias{masonryRow}
\title{Row}
\usage{
masonryRow(..., classes = "")
}
\arguments{
\item{...}{Any number of \code{\link[=masonryItem]{masonryItem()}}.}

\item{classes}{Any additional classes.}
}
\description{
Masonry row
}
